+++ tests/test_serialization.py
@@ -4,6 +4,7 @@ from collections import namedtuple, OrderedDict
 import datetime as dt
 import itertools
 import decimal
+import uuid
 
 import pytest
 
@@ -116,6 +117,27 @@ class TestFieldSerialization:
     def test_callable_field(self, user):
        field = fields.String()
        assert field.serialize('call_me', user) == 'This was called.'
+
+    def test_uuid_field(self, user):
+        user.uuid1 = '{12345678-1234-5678-1234-567812345678}'
+        user.uuid2 = uuid.UUID('12345678123456781234567812345678')
+        user.uuid3 = None
+        user.uuid4 = 'this is not a UUID'
+        user.uuid5 = 42
+        user.uuid6 = {}
+
+        field = fields.UUID()
+        assert isinstance(field.serialize('uuid1', user), str)
+        assert field.serialize('uuid1', user) == '12345678-1234-5678-1234-567812345678'
+        assert isinstance(field.serialize('uuid2', user), str)
+        assert field.serialize('uuid2', user) == '12345678-1234-5678-1234-567812345678'
+        assert field.serialize('uuid3', user) is None
+        with pytest.raises(ValidationError):
+            field.serialize('uuid4', user)
+        with pytest.raises(ValidationError):
+            field.serialize('uuid5', user)
+        with pytest.raises(ValidationError):
+            field.serialize('uuid6', user)
 
     def test_decimal_field(self, user):
         user.m1 = 12
@@ -250,6 +272,20 @@ class TestFieldSerialization:
         assert isinstance(m7s, decimal.Decimal)
         assert m7s.is_zero() and m7s.is_signed()
 
+        field = fields.Decimal(as_string=True, allow_nan=True)
+
+        m2s = field.serialize('m2', user)
+        assert isinstance(m2s, basestring)
+        assert m2s == user.m2
+
+        m5s = field.serialize('m5', user)
+        assert isinstance(m5s, basestring)
+        assert m5s == user.m5
+
+        m6s = field.serialize('m6', user)
+        assert isinstance(m6s, basestring)
+        assert m6s == user.m6
+
     def test_decimal_field_special_values_not_permitted(self, user):
         user.m1 = '-NaN'
         user.m2 = 'NaN'
@@ -277,6 +313,28 @@ class TestFieldSerialization:
         m7s = field.serialize('m7', user)
         assert isinstance(m7s, decimal.Decimal)
         assert m7s.is_zero() and m7s.is_signed()
+
+    def test_decimal_field_fixed_point_representation(self, user):
+        """
+        Test we get fixed-point string representation for a Decimal number that would normally
+        output in engineering notation.
+        """
+        user.m1 = '0.00000000100000000'
+
+        field = fields.Decimal()
+        s = field.serialize('m1', user)
+        assert isinstance(s, decimal.Decimal)
+        assert s == decimal.Decimal('1.00000000E-9')
+
+        field = fields.Decimal(as_string=True)
+        s = field.serialize('m1', user)
+        assert isinstance(s, basestring)
+        assert s == user.m1
+
+        field = fields.Decimal(as_string=True, places=2)
+        s = field.serialize('m1', user)
+        assert isinstance(s, basestring)
+        assert s == '0.00'
 
     def test_boolean_field_serialization(self, user):
         field = fields.Boolean()
@@ -472,7 +530,7 @@ class TestFieldSerialization:
 
     def test_time_field(self, user):
         field = fields.Time()
-        expected = user.time_registered.isoformat()[:12]
+        expected = user.time_registered.isoformat()[:15]
         assert field.serialize('time_registered', user) == expected
 
         user.time_registered = None